requirements.txt
fastapi
uvicorn
pillow
numpy
tflite-runtime==2.11.0